+++ docs/source/changelog.rst
@@ -2,6 +2,15 @@
 Changelog
 =========
 
+1.0.1
+^^^^^
+
+December 1, 2020
+
+- Added lets/launch/elasticsearch
+- Added lets/launch/kibana
+
+
 1.0.0
 ^^^^^